Exploratory data analysis of the Haberman's Cancer survival dataset
Haberman's Survival Dataset contains cases from a study conducted on the survival of patients who had undergone surgery for breast cancer between 1958 and 1970 at the University of Chicago's Billings Hospital. My aim in this study is to conduct some exploratory data analysis (EDA) to find out if there are any patterns in this data and to handle any missing values (if present).